fix: estd_db_time_factor 타입을 int에서 float로 변경
- SGA advice 파싱 시 소수점 값(.4188 등)을 올바르게 처리
- base_models.py: estd_db_time_factor 타입 int -> float
- statspack_parser.py: int() -> float() 변환
- test_dbcsi_data_models_properties.py: 테스트 데이터 생성 수정

+++ src/dbcsi/models/base_models.py
@@ -131,7 +131,7 @@ class SGAAdvice:
     sga_size: int
     sga_size_factor: float
     estd_db_time: int
-    estd_db_time_factor: int
+    estd_db_time_factor: float  # AWR에서 소수점 값 (예: 0.4188)
     estd_physical_reads: int
     sga_target: int
